Almost Finished
I lost access to the AWS lab at 7:30 PM, 4.5 hours before the due date.

The tutorial is accurate assuming I can fix the EULA bug, but I cannot attempt fixing it without access to the AWS lab.

This is mostly functional outside the EULA not working as intended. I have sent an email to the professor asking if I can have access to the AWS lab before midnight so that I may fix my bug and record a video.

For now I am commit-ing my progress in the hopes of at least partial credit (worst case scenario) and not failing the class.